+++ 043_calc_factorial_loop/main.c
@@ -5,7 +5,7 @@
 int main() {
 	// declarations
 	int num, i;
-	long fact = 1;
+	long long fact = 1;
 
 	// input
 	printf("Enter a number: ");
@@ -17,7 +17,7 @@ int main() {
 	}
 
 	// output
-	printf("Factorial = %ld", fact);
+	printf("Factorial = %lld", fact);
 
 	// exit with code 0
 	return 0;
